– Helper –
Working on helper and story today, since we need to outline basic story structure in order to fill it in with actions and code.

Helper interface contains random number generator, as well as the display method.

Probabilities match the game throughout on the scale from 0 to 100, with a step of 10, hence it would comply with the idea of modularity of game's functionality.

Story files still needs to be filled in.